Fix plot updating from script to follow active/hold consistently
